hw3 - add dialog confirmation

+++ src/app/home/home.component.ts
@@ -1,6 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {Course} from "../common/entities/course";
 import {CoursesService} from "./services/courses.service";
+import {CourseDeleteConfirmationComponent} from "./course-delete-confirmation/course-delete-confirmation.component";
+import {MdDialog} from "@angular/material";
 
 @Component({
   selector: 'trainme-home',
@@ -12,12 +14,19 @@ export class HomeComponent implements OnInit {
 
   courses: Array<Course>;
 
-  constructor(private coursesService: CoursesService) {
+  constructor(private coursesService: CoursesService,
+              private dialog: MdDialog) {
   }
 
   courseDeleteHandler(course: Course) {
-   this.coursesService.delete(course)
-     .subscribe(()=>this.courses.slice(this.courses.indexOf(course), 1));
+    let dialogRef = this.dialog.open(CourseDeleteConfirmationComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.coursesService.delete(course)
+          .subscribe(() => this.courses.slice(this.courses.indexOf(course), 1));
+      }
+    });
+
   }
 
   ngOnInit() {
